Move kube and docker clients to shared module
since these are now used in multiple places make em reusable
